Complete Phase 4: State and Context Refactoring
- Finalized the decoupling of navigation logic from the application's state management by removing routing-related state variables from `HabitsStateContext`.
- Updated components (`Header`, `Body`, `HabitsOverview`, and `ProgressOverview`) to operate independently of the removed state flags, ensuring cleaner separation of concerns.
- Implemented legacy navigation warnings in the `Header` and `ProgressOverview` components to maintain visual compatibility while transitioning to URL-based routing.
- Documented the implementation details and benefits in `PHASE-4-IMPLEMENTATION.md`, highlighting improvements in code clarity, maintainability, and testability.
