+++ IMPLEMENTATION-PLAN-React-Router.md
@@ -90,29 +90,33 @@ Engineering Requirements Document.
   - Code examples and architecture decisions
   - Benefits and validation results
 
-## Phase 3: Core Application Integration
+## Phase 3: Core Application Integration ✅ **COMPLETED**
 
-### Task 3.1: Update Application Entry
+### Task 3.1: Update Application Entry ✅
 
-- [ ] Integrate `RouterProvider` in `main.tsx`
-- [ ] Ensure `App.tsx` wraps context providers around `RouterProvider`
+- [x] Integrate `RouterProvider` in `main.tsx`
+- [x] Ensure `App.tsx` wraps context providers around `RouterProvider`
 
-### Task 3.2: Refactor Components
+### Task 3.2: Refactor Components ✅
 
-- [ ] Update `Body.tsx` to use `Outlet`
-- [ ] Modify `Header.tsx` to replace buttons with `NavLink`
+- [x] Update `Body.tsx` to use `Outlet`
+- [x] Modify `Header.tsx` to replace buttons with `NavLink`
 
-## Phase 4: State and Context Refactoring
+## Phase 4: State and Context Refactoring ✅ **COMPLETED**
 
-### Task 4.1: Update HabitsStateContext
+### Task 4.1: Update HabitsStateContext ✅
 
-- [ ] Remove routing-related state variables
+- [x] Remove routing-related state variables
   - `isHabitsOverviewMode`, `isProgressOverviewMode`
   - Associated toggle functions
 
-### Task 4.2: Ensure Independence from Removed States
+### Task 4.2: Ensure Independence from Removed States ✅
 
-- [ ] Refactor components to operate independently of removed state flags
+- [x] Refactor components to operate independently of removed state flags
+  - Updated `Header.tsx` to work without navigation state
+  - Updated `Body.tsx` to default to Daily Habit Tracker in legacy mode
+  - Updated `HabitsOverview.tsx` to use React Router navigation
+  - Updated `ProgressOverview.tsx` to use React Router navigation
 
 ## Phase 5: Testing and Validation
