+++ src/features/habits/components/AddHabitDialog.tsx
@@ -1,7 +1,7 @@
 import HabitDialog from '@/features/habits/components/HabitDialog';
 import HabitDialogHeader from '@/features/habits/components/HabitDialogHeader';
 import HabitForm from '@/features/habits/components/HabitForm';
-import { useHabitsContext } from '@/features/habits/context/HabitsStateContext';
+import { useHabitsContext } from '@/features/habits/hooks/useHabitsContext';
 
 export default function AddHabitDialog() {
   const { showAddHabitDialog, toggleShowAddHabitDialog } = useHabitsContext();
