+++ src/features/habits/components/EditHabitDialog.tsx
@@ -3,7 +3,7 @@ import HabitDialog from '@/features/habits/components/HabitDialog';
 import HabitDialogHeader from '@/features/habits/components/HabitDialogHeader';
 import HabitForm from '@/features/habits/components/HabitForm';
 import HabitStats from '@/features/habits/components/HabitStats';
-import { useHabitsContext } from '@/features/habits/context/HabitsStateContext';
+import { useHabitsContext } from '@/features/habits/hooks/useHabitsContext';
 
 export default function EditHabitDialog() {
   const { editingHabit, showEditHabitDialog, toggleShowEditHabitDialog } =
