+++ src/features/layout/components/Body.tsx
@@ -3,14 +3,9 @@ import { Outlet } from 'react-router-dom';
 
 import Welcome from '@/features/layout/components/Welcome';
 import DailyHabitTracker from '@/features/habits/components/DailyHabitTracker';
-import ProgressOverview from '@/features/progress/components/ProgressOverview';
-import HabitsOverview from '@/features/habits/components/HabitsOverview';
-import { useHabitsContext } from '@/features/habits/context/HabitsStateContext';
 import { featureFlags } from '@/config/featureFlags';
 
 export default function Body() {
-  const { isHabitsOverviewMode, isProgressOverviewMode } = useHabitsContext();
-
   if (featureFlags.isUrlRoutingEnabled) {
     return (
       <>
@@ -25,6 +20,7 @@ export default function Body() {
     );
   }
 
+  // Legacy state-based navigation - default to Daily Habit Tracker
   return (
     <>
       <SignedOut>
@@ -32,22 +28,8 @@ export default function Body() {
         <Welcome />
       </SignedOut>
       <SignedIn>
-        {isHabitsOverviewMode ? (
-          <h1 className="sr-only">Habits Overview</h1>
-        ) : isProgressOverviewMode ? (
-          <h1 className="sr-only">Progress Overview</h1>
-        ) : (
-          <h1 className="sr-only">Daily Habit Tracker</h1>
-        )}
-        {isHabitsOverviewMode ? (
-          <HabitsOverview />
-        ) : isProgressOverviewMode ? (
-          <ProgressOverview />
-        ) : (
-          <>
-            <DailyHabitTracker />
-          </>
-        )}
+        <h1 className="sr-only">Daily Habit Tracker</h1>
+        <DailyHabitTracker />
       </SignedIn>
     </>
   );
